test(moe): consolidate unified FP8 conformance
Merge block-scale and per-tensor TRTLLM FP8 coverage into one architecture-gated suite while keeping format-specific helpers and sections isolated.
